refactor:  JSON 파일을 읽어 도서관 목록을 저장하는 기능 분리
- `LibraryDomainService`의 saveLibraryFromJson 메서드를 JSON 파일을 읽는 메서드와 도서관 목록을 저장하는 메서드로 분리
- JSON 파일을 읽는 메서드를 `JsonConvertor` 클래스로 분리
