+++ be/ebook-search/ebook-crawler/src/main/java/com/meetyourbook/controller/LibraryCrawlerController.java
@@ -12,12 +12,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class LibraryCrawlerController {
 
+    public static final String JSON_FILE_PATH = "/Users/jeonbyeong-ung/IdeaProjects/meet-your-book/be/ebook-search/ebook-crawler/src/main/resources/library_list.json";
     private final LibraryImportService libraryImportService;
 
     @PostMapping("/save")
     public ResponseEntity<?> saveLibraryFromJson() {
-        libraryImportService.saveLibraryFromJson(
-            "/Users/jeonbyeong-ung/IdeaProjects/meet-your-book/be/ebook-search/ebook-crawler/src/main/resources/library_list.json");
+        libraryImportService.saveLibraryFromJson(JSON_FILE_PATH);
         return ResponseEntity.ok().build();
     }
 }
