+++ be/ebook-search/ebook-crawler/src/main/java/com/meetyourbook/service/LibraryImportService.java
@@ -1,5 +1,8 @@
 package com.meetyourbook.service;
 
+import com.meetyourbook.dto.LibraryCreationInfo;
+import com.meetyourbook.util.JsonConvertor;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -7,10 +10,12 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class LibraryImportService {
 
-    private final LibraryDomainService libraryDomainService;
+    private final LibraryDomainService libraryService;
+    private final JsonConvertor jsonConvertor;
 
     public void saveLibraryFromJson(String path) {
-        libraryDomainService.saveLibraryFromJson(path);
+        List<LibraryCreationInfo> libraryCreationInfos = jsonConvertor.readFromJson(path);
+        libraryService.createLibraries(libraryCreationInfos);
     }
 
 }
