+++ be/ebook-search/ebook-domain/src/main/java/com/meetyourbook/service/LibraryDomainService.java
@@ -1,15 +1,11 @@
 package com.meetyourbook.service;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.meetyourbook.dto.LibraryCreationInfo;
 import com.meetyourbook.dto.LibraryResponse;
 import com.meetyourbook.dto.LibraryUpdateInfo;
 import com.meetyourbook.entity.Library;
 import com.meetyourbook.exception.ResourceNotFoundException;
 import com.meetyourbook.repository.LibraryRepository;
-import java.io.File;
-import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
@@ -23,7 +19,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class LibraryDomainService {
 
     private final LibraryRepository libraryRepository;
-    private final ObjectMapper objectMapper;
 
     @Transactional(readOnly = true)
     public List<Library> findAll() {
@@ -34,6 +29,14 @@ public class LibraryDomainService {
     public Long createLibrary(LibraryCreationInfo creationInfo) {
         Library library = libraryRepository.save(creationInfo.toEntity());
         return library.getId();
+    }
+
+    @Transactional
+    public void createLibraries(List<LibraryCreationInfo> libraryCreationInfos) {
+        List<Library> libraries = libraryCreationInfos.stream()
+            .map(LibraryCreationInfo::toEntity)
+            .toList();
+        libraryRepository.saveAll(libraries);
     }
 
     @Transactional
@@ -74,31 +77,6 @@ public class LibraryDomainService {
     public Library findByBaseUrl(String baseUrl) {
         return libraryRepository.findByLibraryUrl_UrlContaining(baseUrl)
             .orElseThrow(() -> new NoSuchElementException("base Url not found = " + baseUrl));
-    }
-
-    @Transactional
-    public void saveLibraryFromJson(String filePath) {
-        try {
-            List<LibraryCreationInfo> libraryCreationInfos = readLibraryCreationsFromJson(filePath);
-            List<Library> libraries = convertToLibraries(libraryCreationInfos);
-            libraryRepository.saveAll(libraries);
-            log.info("Successfully saved {} libraries", libraries.size());
-        } catch (IOException e) {
-            log.error("Error reading JSON file: {}", filePath, e);
-            throw new RuntimeException("Failed to read library data from JSON", e);
-        }
-    }
-
-    private List<LibraryCreationInfo> readLibraryCreationsFromJson(String filePath)
-        throws IOException {
-        return objectMapper.readValue(new File(filePath), new TypeReference<>() {
-        });
-    }
-
-    private List<Library> convertToLibraries(List<LibraryCreationInfo> libraryCreationInfos) {
-        return libraryCreationInfos.stream()
-            .map(LibraryCreationInfo::toEntity)
-            .toList();
     }
 
     private Library findLibrary(Long id) {
